rename cloudinary folders constants file
